+++ src/core/mod.rs
@@ -1,3 +1,4 @@
+pub mod autostart;
 pub mod clipboard;
 pub mod font;
 #[allow(dead_code)]